+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## Unreleased
+## 1.3.1 - 2022-06-13
 
 ### Added
 - Support for session persistence. If a Set-Cookie HTTP header is present the SDK will now set a Cookie header using the requested session value.